+++ st2common/tests/integration/test_register_content_script.py
@@ -100,7 +100,7 @@ class ContentRegisterScripTestCase(IntegrationTestCase):
         exit_code, stdout, stderr = run_command(cmd=cmd)
         self.assertTrue('Registering actions' in stderr)
         self.assertTrue('Registering rules' in stderr)
-        self.assertTrue('Setup virtualenv for 5 pack(s)' in stderr)
+        self.assertTrue('Setup virtualenv for 6 pack(s)' in stderr)
         self.assertEqual(exit_code, 0)
 
     def test_register_setup_virtualenvs(self):
@@ -118,5 +118,5 @@ class ContentRegisterScripTestCase(IntegrationTestCase):
         # All packs
         cmd = BASE_CMD_ARGS + ['--register-setup-virtualenvs', '--register-no-fail-on-failure']
         exit_code, stdout, stderr = run_command(cmd=cmd)
-        self.assertTrue('Setup virtualenv for 5 pack(s)' in stderr)
+        self.assertTrue('Setup virtualenv for 6 pack(s)' in stderr)
         self.assertEqual(exit_code, 0)